Extracted the Pagination component logic into a new usePagination custom hook
