+++ src/common/Pagination/Buttons/FirstPageButton/index.js
@@ -1,5 +1,3 @@
-import { useDispatch } from "react-redux";
-import { resetPage } from "../../paginationSlice";
 import {
     PageVectorIcon,
     ButtonContent,
@@ -7,29 +5,22 @@ import {
 } from "../../styled";
 import { SmallTile } from "../../../styled";
 
-const FirstPageButton = ({ currentPage }) => {
-    const dispatch = useDispatch();
-
-    const goToFirstPage = () => {
-        if (currentPage !== 1) {
-            dispatch(resetPage());
-        }
-    };
-
-    return (
-        <SmallTile
-            onClick={goToFirstPage}
-            $pageButton
-            as="button"
-            disabled={currentPage === 1}
-        >
-            <ButtonContent>
-                <PageVectorIcon $previousPage />
-                <PageVectorIcon $previousPage $mobileOnly />
-                <MobileHide>First</MobileHide>
-            </ButtonContent>
-        </SmallTile>
-    )
-};
+const FirstPageButton = ({
+    currentPage,
+    buttonFunction
+}) => (
+    <SmallTile
+        onClick={buttonFunction}
+        $pageButton
+        as="button"
+        disabled={currentPage === 1}
+    >
+        <ButtonContent>
+            <PageVectorIcon $previousPage />
+            <PageVectorIcon $previousPage $mobileOnly />
+            <MobileHide>First</MobileHide>
+        </ButtonContent>
+    </SmallTile>
+);
 
 export default FirstPageButton;
