+++ src/common/Pagination/Buttons/NextPageButton/index.js
@@ -1,5 +1,3 @@
-import { useDispatch } from "react-redux";
-import { changePage } from "../../paginationSlice";
 import {
     PageVectorIcon,
     ButtonContent,
@@ -9,29 +7,20 @@ import { SmallTile } from "../../../styled";
 
 const NextPageButton = ({
     totalPages,
-    currentPage
-}) => {
-    const dispatch = useDispatch();
-
-    const goToNextPage = () => {
-        if (currentPage < totalPages) {
-            dispatch(changePage({ payload: currentPage + 1 }));
-        }
-    };
-
-    return (
-        <SmallTile
-            onClick={goToNextPage}
-            $pageButton
-            as="button"
-            disabled={currentPage === totalPages}
-        >
-            <ButtonContent>
-                <MobileHide>Next</MobileHide>
-                <PageVectorIcon />
-            </ButtonContent>
-        </SmallTile>
-    )
-};
+    currentPage,
+    buttonFunction
+}) => (
+    <SmallTile
+        onClick={buttonFunction}
+        $pageButton
+        as="button"
+        disabled={currentPage === totalPages}
+    >
+        <ButtonContent>
+            <MobileHide>Next</MobileHide>
+            <PageVectorIcon />
+        </ButtonContent>
+    </SmallTile>
+);
 
 export default NextPageButton;
